Adds psbt creation and signing support in JM wallet.
This is a first draft and lacks detail. Basically:
Subclassed btcointx.CKeyBase to create a privkey type
that uses our signing code.
Sublcassed bitcointx.KeyStore to create a very simple
KeyStore consisting of lists of the above keys.
These are then used to allow signing of inputs in
transactions owned by our wallet.
Created a PSBTWalletMixin to perform the above function,
and added it to our three wallet types (although only
the default "SegwitLegacy" is tested for now).
Wrote a small proof of concept in the direct_send()
function, which creates a psbt for the transaction, then
validates and signs it. This POC will be removed though,
for more useful code.