Address requirements of python-bitcointx 1.1.0
Specifically, the witness `utxo` field can no longer be
assumed to be of type CTxOut, so we should access the
CTxOut with the field witness_utxo and also when updating
the `utxo` field we now use `set_utxo()`.
Additionally some minor typos/comment corrections and removal
of the now defunct `apply_freeze_signature`.